Theme modification
- Add Yellow Theme
- Change the name of some themes
