+++ lib/l10n/app_zh.arb
@@ -107,11 +107,12 @@
   "themeListTitle": "主題列表",
   "themeListDescription": "選擇一個主題",
   "themeListNameDefault": "預設",
-  "themeListNameReverse": "反轉",
+  "themeListNameMaterial": "質感",
+  "themeListNameYellow": "黃色",
 
   "brightnessSelectorTitle": "亮度",
   "brightnessSelectorDescription": "選擇亮度",
   "brightnessSystem": "系統",
-  "brightnessLight": "明亮",
-  "brightnessDark": "黑暗"
+  "brightnessLight": "淺色",
+  "brightnessDark": "深色"
 }
